Reworked the OP codes; Can now parse all OP codes into Instructions
The OP codes are now parsed into a convenient Instruction struct.

I'm not too happy about a couple of design decisions already, such as how the
registers have been implmented (both in the CPU and the enum that's used for the
Instructions) so maybe I can come back and fix it up later. The decode function
looks a bit crazy, but at least it's dead simple to follow, which will help
when debugging.