Moved all main function to a new chip8.c file
The main file used to live in the framebuffer.c file, but that was just to get
things started. All the code in framebuffer.c and framebuffer.h should really
just be things related to the framebuffer.

framebuffer.c and framebuffer.h have been cleaned up.

I still need to figure out why I can't move the #if DEBUG to "chip8.h".
Currently I get errors when I try to move that, despite the fact that chip8.h is
the first include in chip8.c.