Moved draw_buffer to the main file.
Additionally, the color setup for drawing a pixel to the screen has been added
to chip8.c. After each fetch --> decode --> execute cycle, draw_buffer is
being called.

I could add a flag on the CPU that would signal if a change has occurred on the
framebuffer. That way the draw_buffer and refresh would only occur if the this
flag has been set on the CPU.

Finally, I'm not sure if I like the framebuffer as a char array. Since
according to the CHIP8 spec the framebuffer only stores information on whether
or not the pixel is on, I'm using 8 bits to store 1 bit of information (on or
off).

So it might be better to have the framebuffer just be an array of 8x4 bytes
(64 for bits for the height and 32 bits for the width). It would make things
more "economical" but it would require reworking some of the existing
framebuffer code.