Reworked debug mode to be played alongside regular mode.
The debug window will now be displayed next to the regular mode window when the
program is called with debug mode on.

The debug window no longer waits for the user to step through the program. The
program executes normally and the debug output is now shown to the right of the
regular execution of the program.