fix(keypad): improve keypress responsiveness
Originally, the keypress reading was very slow and unresponsive. This
patch adds a keypad that listens for keypresses in a separate thread. It
keeps track of what keys have been pressed and manages a timer for
unsetting the keypresses.

A keypad for visualizing the keypresses has also been added to the debug
window.