feat: simulate running the emulator at 500 Hz
Previously, we were just sleeping for 2000 microseconds after each
cycle. Reading around online, it seems that the emulator is supposed to
run at 500 Hz.

This feature adds the calculations to run at emulator at 500 Hz. Things
such as the decrementing of the timers and drawing are only done now
60 times a second, whereas previously, they were incorrectly being done
after every cycle.